+++ services/frontend-service/src/legacy-ui/ActionDetails.tsx
@@ -27,7 +27,6 @@ export enum ActionTypes {
     Deploy,
     PrepareUndeploy,
     Undeploy,
-    DeleteQueue,
     CreateEnvironmentLock,
     DeleteEnvironmentLock,
     CreateApplicationLock,
@@ -52,14 +51,6 @@ export type PrepareUndeploy = {
 export type Undeploy = {
     undeploy: {
         application: string;
-    };
-};
-
-export type DeleteQueue = {
-    deleteQueue: {
-        environment: string;
-        application: string;
-        currentlyDeployedVersion: number;
     };
 };
 
@@ -97,7 +88,6 @@ export type CartAction =
     | Deploy
     | PrepareUndeploy
     | Undeploy
-    | DeleteQueue
     | CreateEnvironmentLock
     | DeleteEnvironmentLock
     | CreateApplicationLock
@@ -108,12 +98,7 @@ type lockAction = typeof lockActions[number];
 export const hasLockAction = (actions: CartAction[]) =>
     actions.some((act) => lockActions.includes(getActionDetails(act).type as lockAction));
 
-const deployActions = [
-    ActionTypes.Deploy,
-    ActionTypes.PrepareUndeploy,
-    ActionTypes.Undeploy,
-    ActionTypes.DeleteQueue,
-] as const;
+const deployActions = [ActionTypes.Deploy, ActionTypes.PrepareUndeploy, ActionTypes.Undeploy] as const;
 type deployAction = typeof deployActions[number];
 export const isDeployAction = (act: CartAction) => deployActions.includes(getActionDetails(act).type as deployAction);
 
@@ -220,24 +205,6 @@ export const getActionDetails = (action: CartAction): ActionDetails => {
             summary: 'Undeploy and delete Application ' + action.undeploy.application,
             icon: <DeleteForeverRounded />,
             application: action.undeploy.application,
-        };
-    else if ('deleteQueue' in action)
-        return {
-            type: ActionTypes.DeleteQueue,
-            name: 'Delete Queue',
-            dialogTitle: 'Please be aware:',
-            description:
-                'This action deletes the queued releases for the service "' +
-                action.deleteQueue.application +
-                '" on ' +
-                action.deleteQueue.environment +
-                '.Technically, it deploys the version that is already deployed(' +
-                action.deleteQueue.currentlyDeployedVersion +
-                '), which as a side effect deletes the queue.',
-            summary: 'Delete queue for "' + action.deleteQueue.application + '" on ' + action.deleteQueue.environment,
-            icon: <MoveToInboxRounded />,
-            environment: action.deleteQueue.environment,
-            application: action.deleteQueue.application,
         };
     else
         return {
@@ -358,20 +325,6 @@ export const transformToBatch = (act: CartAction): BatchAction | null => {
                 $case: 'undeploy',
                 undeploy: {
                     application: act.undeploy.application,
-                },
-            },
-        };
-    else if ('deleteQueue' in act)
-        // Deleting the queue is just deploying the currently deployed version again
-        return {
-            action: {
-                $case: 'deploy',
-                deploy: {
-                    environment: act.deleteQueue.environment,
-                    application: act.deleteQueue.application,
-                    version: act.deleteQueue.currentlyDeployedVersion,
-                    lockBehavior: LockBehavior.Ignore,
-                    ignoreAllLocks: false,
                 },
             },
         };